+++ cfripper/rules/IAMManagedPolicyWildcardActionRule.py
@@ -12,6 +12,8 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
+from pycfmodel.model.resources.iam_managed_policy import IAMManagedPolicy
+
 from ..config.regex import REGEX_WILDCARD_POLICY_ACTION
 from ..model.rule import Rule
 
@@ -22,7 +24,7 @@ class IAMManagedPolicyWildcardActionRule(Rule):
 
     def invoke(self, cfmodel):
         for logical_id, resource in cfmodel.Resources.items():
-            if resource.Type == "AWS::IAM::ManagedPolicy" and resource.Properties.PolicyDocument.allowed_actions_with(
+            if isinstance(resource, IAMManagedPolicy) and resource.Properties.PolicyDocument.allowed_actions_with(
                 REGEX_WILDCARD_POLICY_ACTION
             ):
                 self.add_failure(type(self).__name__, self.REASON.format(logical_id))