+++ cfripper/rules/SQSQueuePolicyNotPrincipalRule.py
@@ -12,6 +12,8 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
+from pycfmodel.model.resources.sqs_queue_policy import SQSQueuePolicy
+
 from ..model.enums import RuleMode
 from ..model.rule import Rule
 
@@ -22,7 +24,7 @@ class SQSQueuePolicyNotPrincipalRule(Rule):
 
     def invoke(self, cfmodel):
         for logical_id, resource in cfmodel.Resources.items():
-            if resource.Type == "AWS::SQS::QueuePolicy":
+            if isinstance(resource, SQSQueuePolicy):
                 for statement in resource.Properties.PolicyDocument._statement_as_list():
                     if statement.NotPrincipal:
                         self.add_failure(type(self).__name__, self.REASON.format(logical_id))