+++ cfripper/rules/SecurityGroupIngressOpenToWorld.py
@@ -12,13 +12,15 @@ under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
+from pycfmodel.model.resources.security_group_ingress import SecurityGroupIngress
+
 from .SecurityGroupOpenToWorldRule import SecurityGroupOpenToWorldRule
 
 
 class SecurityGroupIngressOpenToWorld(SecurityGroupOpenToWorldRule):
     def invoke(self, cfmodel):
         for logical_id, resource in cfmodel.Resources.items():
-            if resource.Type == "AWS::EC2::SecurityGroupIngress" and (
+            if isinstance(resource, SecurityGroupIngress) and (
                 resource.ipv4_slash_zero() or resource.ipv6_slash_zero()
             ):
                 for port in range(resource.Properties.FromPort, resource.Properties.ToPort + 1):